git-svn-id: https://gamescrafters.svn.sourceforge.net/svnroot/gamescrafters/GamesmanJava@5918 a668be5d-ef11-cb4c-8bfd-e08cbb24f2ba

+++ hadoop-src/edu/berkeley/gamesman/parallel/tier/HadoopTierMaster.java
@@ -74,6 +74,8 @@ public class HadoopTierMaster implements Runnable {
 		do {
 			try {
 				success = job.waitForCompletion(true);
+				if (!success)
+					throw new Error("Tier " + tier + " failed");
 				fs.delete(outputDirectory, true);
 				System.out
 						.println("On tier " + tier + ", success = " + success);
